+++ docs/examples/samples/relight.ts
@@ -30,7 +30,9 @@ addEventListener('resize', () => {
   renderer.setSize(innerWidth, innerHeight);
 });
 
-const splats = await StreamedSplatMesh.load('/remote/jack/v/Dehaar/Dehaar.lcc2');
+const splats = await StreamedSplatMesh.load(
+  'https://assets.voluma.ai/voluma/cultural-heritage/Tempel/Tempel.lcc2',
+);
 scene.add(splats);
 splats.updateWorldMatrix(true, false);